+++ aitune/torch/backend/tensorrt/__init__.py
@@ -18,11 +18,17 @@ TensorRT engines for accelerated inference. It includes components for ONNX
 export, TensorRT engine building, and runtime inference.
 """
 
+from aitune.torch.backend.tensorrt.onnx_autocast import ONNXAutoCastConfig
+from aitune.torch.backend.tensorrt.onnx_quantization import ONNXQuantizationConfig
 from aitune.torch.backend.tensorrt.tensorrt_backend import TensorRTBackend, TensorRTBackendConfig
 from aitune.torch.backend.tensorrt.tensorrt_profile import TensorRTProfile
+from aitune.torch.backend.tensorrt.torch_quantization import TorchQuantizationConfig
 
 __all__ = [
     "TensorRTBackend",
     "TensorRTBackendConfig",
     "TensorRTProfile",
+    "ONNXAutoCastConfig",
+    "ONNXQuantizationConfig",
+    "TorchQuantizationConfig",
 ]